Use the current thread's context ClassLoader
This ensures that the runtime Mirror passed to the TypeCreator will
match the used to create the Type. When running from the Spark REPL, it
did not previouly match.

This may not yet be thread-safe (i.e. when running across multiple Spark
Executors in the same process), as TypeTags are cached across threads in
EncoderCache.